+++ src/main/java/com/example/restfulwebservice/helloworld/HelloWorldBean.java
@@ -1,4 +1,4 @@
-package com.example.restfulwebservice;
+package com.example.restfulwebservice.helloworld;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
